requirements.txt
altair==4.2.2
attrs==23.1.0
azure-ai-textanalytics==5.2.1
azure-cognitiveservices-speech==1.27.0
azure-common==1.1.28
azure-core==1.26.4
backports.zoneinfo==0.2.1
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
install==1.3.5
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
msrest==0.7.1
murmurhash==1.0.9
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
pathy==0.10.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
preshed==3.0.8
protobuf==3.20.3
pyarrow==11.0.0
pydantic==1.10.7
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.29.0
requests-oauthlib==1.3.1
rich==13.3.4
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
streamlit==1.21.0
thinc==8.1.9
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
typer==0.7.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
wasabi==1.1.1
zipp==3.15.0
